+++ jest.config.js
@@ -3,6 +3,7 @@
 module.exports = {
   testEnvironment: 'node',
   collectCoverage: true,
+  testEnvironment: 'node',
   coverageReporters: ['text', 'html'],
   setupFilesAfterEnv: ['jest-extended'],
   testMatch: ['**/test/**/*.[jt]s?(x)']